Reorganize imports for tabs.
