+++ src/tabs/index.ts
@@ -1,0 +1,10 @@
+export { default as AutoCorrelationTab } from './AutoCorrelationTab'
+export { default as ConnectionTab } from './ConnectionTab'
+export { default as Diagnostics } from './DiagnosticsTab'
+export { default as ExportTab } from './ExportTab'
+export { default as HistogramTab } from './HistogramTab'
+export { default as RunInfoTab } from './RunInfoTab'
+export { default as ScatterplotsTab } from './ScatterplotsTab'
+export { default as TabWidget } from './TabWidget/TabWidget'
+export { default as TablesTab } from './TablesTab/TablesTab'
+
